Move GetInputs to its own file
